+++ .pre-commit-config.yaml
@@ -29,6 +29,8 @@ repos:
   rev: 24.4.2
   hooks:
   - id: black
+    args:
+    - --line-length=100
 - repo: https://github.com/PyCQA/autoflake
   rev: v2.3.1
   hooks:
