+++ termspark/painter/constants/fore.py
@@ -1,21 +1,7 @@
 from typing import Final
 
+from termspark.painter.constants.color import Color
 
-class Fore:
-    BLACK: Final[int] = 30
-    RED: Final[int] = 31
-    GREEN: Final[int] = 32
-    YELLOW: Final[int] = 33
-    BLUE: Final[int] = 34
-    MAGENTA: Final[int] = 35
-    CYAN: Final[int] = 36
-    WHITE: Final[int] = 37
-    GRAY: Final[int] = 90
-    GREY: Final[int] = 90
-    LIGHT_RED: Final[int] = 91
-    LIGHT_GREEN: Final[int] = 92
-    LIGHT_YELLOW: Final[int] = 93
-    LIGHT_BLUE: Final[int] = 94
-    LIGHT_MAGENTA: Final[int] = 95
-    LIGHT_CYAN: Final[int] = 96
-    LIGHT_WHITE: Final[int] = 97
+
+class Fore(Color):
+    PREFIX: Final[str] = "\x1b[38;5;"
