+++ termspark/painter/constants/highlight.py
@@ -1,258 +1,7 @@
 from typing import Final
 
+from termspark.painter.constants.color import Color
 
-class Highlight:
-    BLACK: Final[int] = 0
-    MAROON: Final[int] = 1
-    GREEN: Final[int] = 2
-    OLIVE: Final[int] = 3
-    NAVY: Final[int] = 4
-    PURPLE: Final[int] = 5
-    TEAL_1: Final[int] = 6
-    SILVER: Final[int] = 7
-    GRAY: Final[int] = 8
-    GREY: Final[int] = 8
-    RED: Final[int] = 9
-    LIME: Final[int] = 10
-    YELLOW: Final[int] = 11
-    BLUE: Final[int] = 12
-    FUCHSIA: Final[int] = 13
-    AQUA: Final[int] = 14
-    WHITE: Final[int] = 15
-    NAVY_BLUE: Final[int] = 17
-    DARK_BLUE: Final[int] = 18
-    DARK_BLUE_2: Final[int] = 19
-    DARK_BLUE_1: Final[int] = 20
-    BLUE_1: Final[int] = 21
-    DARK_GREEN: Final[int] = 22
-    BLUE_STONE: Final[int] = 23
-    ORIENT: Final[int] = 24
-    ENDEAVOUR: Final[int] = 25
-    SCIENCE_BLUE: Final[int] = 26
-    BLUE_RIBBON: Final[int] = 27
-    JAPANESE_LAUREL: Final[int] = 28
-    DEEP_SEA: Final[int] = 29
-    TEAL: Final[int] = 30
-    DEEP_CERULEAN: Final[int] = 31
-    LOCHMARA: Final[int] = 32
-    AZURE_RADIANCE: Final[int] = 33
-    ISLAMIC_GREEN: Final[int] = 34
-    SPRING_GREEN: Final[int] = 35
-    DARK_CYAN: Final[int] = 36
-    LIGHT_SEA_GREEN: Final[int] = 37
-    CERULEAN: Final[int] = 38
-    BLUE_BOLT: Final[int] = 39
-    ELECTRIC_GREEN: Final[int] = 40
-    MALACHITE: Final[int] = 41
-    CARIBBEAN_GREEN: Final[int] = 42
-    CYAN_1: Final[int] = 43
-    DARK_TURQUOISE: Final[int] = 44
-    VIVID_SKY_BLUE: Final[int] = 45
-    ELECTRIC_GREEN_1: Final[int] = 46
-    GUPPIE_GREEN: Final[int] = 47
-    SPRING_GREEN_1: Final[int] = 48
-    MEDIUM_SPRING_GREEN: Final[int] = 49
-    SEA_GREEN: Final[int] = 50
-    CYAN: Final[int] = 51
-    ROSEWOOD: Final[int] = 52
-    POMPADOUR: Final[int] = 53
-    PIGMENT_INDIGO: Final[int] = 54
-    PURPLE_3: Final[int] = 55
-    ELECTIC_VIOLET: Final[int] = 56
-    BLUE_VIOLET: Final[int] = 57
-    VERDUN_GREEN: Final[int] = 58
-    SCORPION: Final[int] = 59
-    COMET: Final[int] = 60
-    SCAMPI: Final[int] = 61
-    INDIGO: Final[int] = 62
-    CORNFLOWER_BLUE_1: Final[int] = 63
-    LIMEADE: Final[int] = 64
-    GLADE_GREEN: Final[int] = 65
-    JUNIPER: Final[int] = 66
-    HIPPIE_BLUE: Final[int] = 67
-    HAVELOCK_BLUE: Final[int] = 68
-    CORNFLOWER_BLUE: Final[int] = 69
-    LIMEA: Final[int] = 70
-    FERN: Final[int] = 71
-    SILVER_TREE: Final[int] = 72
-    TRADEWIND: Final[int] = 73
-    SHAKESPEARE: Final[int] = 74
-    MALIBU: Final[int] = 75
-    BRIGHT_GREEN: Final[int] = 76
-    PALE_GREEN: Final[int] = 77
-    PASTEL_GREEN: Final[int] = 78
-    DOWNY: Final[int] = 79
-    VIKING: Final[int] = 80
-    STEEL_BLUE: Final[int] = 81
-    CHARTREUSE: Final[int] = 82
-    SCREAMING_GREEN: Final[int] = 83
-    SEA_GREEN_1: Final[int] = 84
-    AQUAMARINE_1: Final[int] = 85
-    AQUAMARINE_2: Final[int] = 86
-    AQUAMARINE: Final[int] = 87
-    DARK_RED: Final[int] = 88
-    FRESH_EGGPLANT: Final[int] = 89
-    DARK_MAGENTA: Final[int] = 90
-    PURPLE_2: Final[int] = 91
-    ELECTRIC_VIOLET: Final[int] = 92
-    PURPLE_1: Final[int] = 93
-    BROWN: Final[int] = 94
-    COPPER_ROSE: Final[int] = 95
-    STRIKE_MASTER: Final[int] = 96
-    DELUGE: Final[int] = 97
-    MEDIUM_PURPLE: Final[int] = 98
-    HELIOTROPE: Final[int] = 99
-    OLIVE_1: Final[int] = 100
-    CLAY_CREEK: Final[int] = 101
-    GRAY_1: Final[int] = 102
-    GREY_1: Final[int] = 102
-    WILD_BLUE_YONDER: Final[int] = 103
-    CHETWODE_BLUE: Final[int] = 104
-    LIGHT_SLATE_BLUE: Final[int] = 105
-    LIMEADE_1: Final[int] = 106
-    CHELSEA_CUCUMBER: Final[int] = 107
-    BAY_LEAF: Final[int] = 108
-    GULF_STREAM: Final[int] = 109
-    POLO_BLUE: Final[int] = 110
-    MALIBU_1: Final[int] = 111
-    PISTACHIO: Final[int] = 112
-    DARK_OLIVE_GREEN: Final[int] = 113
-    FEIJOA: Final[int] = 114
-    VISTA_BLUE: Final[int] = 115
-    BERMUDA: Final[int] = 116
-    ANAKIWA: Final[int] = 117
-    CHARTREUSE_1: Final[int] = 118
-    LIGHT_GREEN: Final[int] = 119
-    MINT_GREEN: Final[int] = 120
-    PALE_GREEN_1: Final[int] = 121
-    AQUA_MARINE: Final[int] = 122
-    ANAKIWA_1: Final[int] = 123
-    BRIGHT_RED: Final[int] = 124
-    FLIRT: Final[int] = 125
-    MEDIUM_VIOLET_RED: Final[int] = 126
-    MAGENTA_1: Final[int] = 127
-    DARK_VIOLET: Final[int] = 128
-    PURPLE_4: Final[int] = 129
-    ROSE_OF_SHARON: Final[int] = 130
-    INDIAN_RED: Final[int] = 131
-    TAPESTRY: Final[int] = 132
-    FUCHSIA_PINK: Final[int] = 133
-    MEDIUM_PURPLE_1: Final[int] = 134
-    HELIOTROPE_1: Final[int] = 135
-    PIRATE_GOLD: Final[int] = 136
-    MUESLI: Final[int] = 137
-    PHARLAP: Final[int] = 138
-    BOUQUET: Final[int] = 139
-    LAVENDER: Final[int] = 140
-    HELIOTROPE_2: Final[int] = 141
-    GOLD_1: Final[int] = 142
-    OLIVE_GREEN: Final[int] = 143
-    HILLARY: Final[int] = 144
-    SILVER_CHALICE: Final[int] = 145
-    WISTFUL: Final[int] = 146
-    MELROSE: Final[int] = 147
-    RIO_GRANDE: Final[int] = 148
-    CONIFER: Final[int] = 149
-    FEIJOA_1: Final[int] = 150
-    PIXIE_GREEN: Final[int] = 151
-    JUNGLE_MIST: Final[int] = 152
-    ANAKIWA_2: Final[int] = 153
-    LIME_1: Final[int] = 154
-    GREEN_YELLOW: Final[int] = 155
-    MINT_GREEN_1: Final[int] = 156
-    DARK_SEA_GREEN: Final[int] = 157
-    AERO_BLUE: Final[int] = 158
-    FRENCH_PASS: Final[int] = 159
-    GUARDSMAN_RED: Final[int] = 160
-    RAZZMATAZZ: Final[int] = 161
-    HOLLYWOOD_CERISE: Final[int] = 162
-    HOLLYWOOD_CERISE_1: Final[int] = 163
-    PURPLE_PIZZAZZ: Final[int] = 164
-    ELECTRIC_VIOLET_1: Final[int] = 165
-    TENN: Final[int] = 166
-    ROMAN: Final[int] = 167
-    CRANBERRY: Final[int] = 168
-    HOPBUSH: Final[int] = 169
-    ORCHID: Final[int] = 170
-    MEDIUM_ORCHID: Final[int] = 171
-    MANGO_TANGO: Final[int] = 172
-    COPPERFIELD: Final[int] = 173
-    PINK: Final[int] = 174
-    CANCAN: Final[int] = 175
-    LIGHT_ORCHID: Final[int] = 176
-    HELIOTROPE_3: Final[int] = 177
-    CORN: Final[int] = 178
-    TACHA: Final[int] = 179
-    TAN: Final[int] = 180
-    CLAM_SHELL: Final[int] = 181
-    THISTLE: Final[int] = 182
-    MAUVE: Final[int] = 183
-    CORN_1: Final[int] = 184
-    KHAKI: Final[int] = 185
-    DECO: Final[int] = 186
-    GREEN_MIST: Final[int] = 187
-    ALTO: Final[int] = 188
-    FOG: Final[int] = 189
-    CHARTREUSE_YELLOW: Final[int] = 190
-    CANARY: Final[int] = 191
-    HONEYSUCKLE: Final[int] = 192
-    REEF: Final[int] = 193
-    SNOWY_MINT: Final[int] = 194
-    OYSTER_BAY: Final[int] = 195
-    ROSE: Final[int] = 197
-    DEEP_PINK: Final[int] = 198
-    HOLLYWOOD_CERISE_2: Final[int] = 199
-    PURPLE_PIZZAZZ_1: Final[int] = 200
-    MAGENTA: Final[int] = 201
-    BLAZE_ORANGE: Final[int] = 202
-    BITTER_SWEET: Final[int] = 203
-    WILD_WATERMELON: Final[int] = 204
-    HOTPINK: Final[int] = 205
-    HOTPINK_1: Final[int] = 206
-    PINK_FLAMINGO: Final[int] = 207
-    FLUSH_ORANGE: Final[int] = 208
-    SALMON: Final[int] = 209
-    VIVID_TANGERINE: Final[int] = 210
-    PINK_SALMON: Final[int] = 211
-    LAVENDER_ROSE: Final[int] = 212
-    BLUSH_PINK: Final[int] = 213
-    YELLOW_SEA: Final[int] = 214
-    TEXAS_ROSE: Final[int] = 215
-    HIT_PINK: Final[int] = 216
-    SUNDOWN: Final[int] = 217
-    COTTON_CANDY: Final[int] = 218
-    LAVENDER_ROSE_1: Final[int] = 219
-    GOLD: Final[int] = 220
-    DANDELION: Final[int] = 221
-    GRANDIS: Final[int] = 222
-    CARAMEL: Final[int] = 223
-    COSMOS: Final[int] = 224
-    PINK_LACE: Final[int] = 225
-    LASER_LEMON: Final[int] = 227
-    DOLLY: Final[int] = 228
-    PORTAFINO: Final[int] = 229
-    CUMULUS: Final[int] = 230
-    COD_GRAY: Final[int] = 232
-    COD_GRAY_1: Final[int] = 233
-    COD_GRAY_2: Final[int] = 234
-    MINE_SHAFT: Final[int] = 235
-    MINE_SHAFT_1: Final[int] = 236
-    MINE_SHAFT_2: Final[int] = 237
-    TUNDORA: Final[int] = 238
-    TUNDORA_1: Final[int] = 239
-    SCORPION_1: Final[int] = 240
-    DOVE_GRAY: Final[int] = 241
-    DOVE_GRAY_1: Final[int] = 242
-    BOULDER: Final[int] = 243
-    GRAY_2: Final[int] = 245
-    GREY_2: Final[int] = 245
-    DUSTY_GRAY: Final[int] = 246
-    SILVER_CHALICE_1: Final[int] = 247
-    SILVER_CHALICE_2: Final[int] = 248
-    SILVER_CHALICE_3: Final[int] = 249
-    SILVER_1: Final[int] = 250
-    SILVER_2: Final[int] = 251
-    ALTO_1: Final[int] = 252
-    ALTO_2: Final[int] = 253
-    MERCURY: Final[int] = 254
-    GALLERY: Final[int] = 255
+
+class Highlight(Color):
+    PREFIX: Final[str] = "\x1b[48;5;"
