+++ tests/color_test.py
@@ -1,0 +1,765 @@
+from termspark.painter.constants.color import Color
+
+
+class TestColor:
+    def test_black(self):
+        assert Color.BLACK == 0
+
+    def test_maroon(self):
+        assert Color.MAROON == 1
+
+    def test_green(self):
+        assert Color.GREEN == 2
+
+    def test_olive(self):
+        assert Color.OLIVE == 3
+
+    def test_navy(self):
+        assert Color.NAVY == 4
+
+    def test_purple(self):
+        assert Color.PURPLE == 5
+
+    def test_teal_1(self):
+        assert Color.TEAL_1 == 6
+
+    def test_silver(self):
+        assert Color.SILVER == 7
+
+    def test_gray(self):
+        assert Color.GRAY == 8
+
+    def test_grey(self):
+        assert Color.GREY == 8
+
+    def test_red(self):
+        assert Color.RED == 9
+
+    def test_lime(self):
+        assert Color.LIME == 10
+
+    def test_yellow(self):
+        assert Color.YELLOW == 11
+
+    def test_blue(self):
+        assert Color.BLUE == 12
+
+    def test_fuchsia(self):
+        assert Color.FUCHSIA == 13
+
+    def test_aqua(self):
+        assert Color.AQUA == 14
+
+    def test_white(self):
+        assert Color.WHITE == 15
+
+    def test_navy_blue(self):
+        assert Color.NAVY_BLUE == 17
+
+    def test_dark_blue(self):
+        assert Color.DARK_BLUE == 18
+
+    def test_dark_blue_2(self):
+        assert Color.DARK_BLUE_2 == 19
+
+    def test_dark_blue_1(self):
+        assert Color.DARK_BLUE_1 == 20
+
+    def test_blue_1(self):
+        assert Color.BLUE_1 == 21
+
+    def test_dark_green(self):
+        assert Color.DARK_GREEN == 22
+
+    def test_blue_stone(self):
+        assert Color.BLUE_STONE == 23
+
+    def test_orient(self):
+        assert Color.ORIENT == 24
+
+    def test_endeavour(self):
+        assert Color.ENDEAVOUR == 25
+
+    def test_science_blue(self):
+        assert Color.SCIENCE_BLUE == 26
+
+    def test_blue_ribbon(self):
+        assert Color.BLUE_RIBBON == 27
+
+    def test_japanese_laurel(self):
+        assert Color.JAPANESE_LAUREL == 28
+
+    def test_deep_sea(self):
+        assert Color.DEEP_SEA == 29
+
+    def test_teal(self):
+        assert Color.TEAL == 30
+
+    def test_deep_cerulean(self):
+        assert Color.DEEP_CERULEAN == 31
+
+    def test_lochmara(self):
+        assert Color.LOCHMARA == 32
+
+    def test_azure_radiance(self):
+        assert Color.AZURE_RADIANCE == 33
+
+    def test_islamic_green(self):
+        assert Color.ISLAMIC_GREEN == 34
+
+    def test_spring_green(self):
+        assert Color.SPRING_GREEN == 35
+
+    def test_dark_cyan(self):
+        assert Color.DARK_CYAN == 36
+
+    def test_light_sea_green(self):
+        assert Color.LIGHT_SEA_GREEN == 37
+
+    def test_cerulean(self):
+        assert Color.CERULEAN == 38
+
+    def test_blue_bolt(self):
+        assert Color.BLUE_BOLT == 39
+
+    def test_electric_green(self):
+        assert Color.ELECTRIC_GREEN == 40
+
+    def test_malachite(self):
+        assert Color.MALACHITE == 41
+
+    def test_caribbean_green(self):
+        assert Color.CARIBBEAN_GREEN == 42
+
+    def test_cyan_1(self):
+        assert Color.CYAN_1 == 43
+
+    def test_dark_turquoise(self):
+        assert Color.DARK_TURQUOISE == 44
+
+    def test_vivid_sky_blue(self):
+        assert Color.VIVID_SKY_BLUE == 45
+
+    def test_electric_green_1(self):
+        assert Color.ELECTRIC_GREEN_1 == 46
+
+    def test_guppie_green(self):
+        assert Color.GUPPIE_GREEN == 47
+
+    def test_spring_green_1(self):
+        assert Color.SPRING_GREEN_1 == 48
+
+    def test_medium_spring_green(self):
+        assert Color.MEDIUM_SPRING_GREEN == 49
+
+    def test_sea_green(self):
+        assert Color.SEA_GREEN == 50
+
+    def test_cyan(self):
+        assert Color.CYAN == 51
+
+    def test_rosewood(self):
+        assert Color.ROSEWOOD == 52
+
+    def test_pompadour(self):
+        assert Color.POMPADOUR == 53
+
+    def test_pigment_indigo(self):
+        assert Color.PIGMENT_INDIGO == 54
+
+    def test_purple_3(self):
+        assert Color.PURPLE_3 == 55
+
+    def test_electic_violet(self):
+        assert Color.ELECTIC_VIOLET == 56
+
+    def test_blue_violet(self):
+        assert Color.BLUE_VIOLET == 57
+
+    def test_verdun_green(self):
+        assert Color.VERDUN_GREEN == 58
+
+    def test_scorpion(self):
+        assert Color.SCORPION == 59
+
+    def test_comet(self):
+        assert Color.COMET == 60
+
+    def test_scampi(self):
+        assert Color.SCAMPI == 61
+
+    def test_indigo(self):
+        assert Color.INDIGO == 62
+
+    def test_cornflower_blue_1(self):
+        assert Color.CORNFLOWER_BLUE_1 == 63
+
+    def test_limeade(self):
+        assert Color.LIMEADE == 64
+
+    def test_glade_green(self):
+        assert Color.GLADE_GREEN == 65
+
+    def test_juniper(self):
+        assert Color.JUNIPER == 66
+
+    def test_hippie_blue(self):
+        assert Color.HIPPIE_BLUE == 67
+
+    def test_havelock_blue(self):
+        assert Color.HAVELOCK_BLUE == 68
+
+    def test_cornflower_blue(self):
+        assert Color.CORNFLOWER_BLUE == 69
+
+    def test_limea(self):
+        assert Color.LIMEA == 70
+
+    def test_fern(self):
+        assert Color.FERN == 71
+
+    def test_silver_tree(self):
+        assert Color.SILVER_TREE == 72
+
+    def test_tradewind(self):
+        assert Color.TRADEWIND == 73
+
+    def test_shakespeare(self):
+        assert Color.SHAKESPEARE == 74
+
+    def test_malibu(self):
+        assert Color.MALIBU == 75
+
+    def test_bright_green(self):
+        assert Color.BRIGHT_GREEN == 76
+
+    def test_pale_green(self):
+        assert Color.PALE_GREEN == 77
+
+    def test_pastel_green(self):
+        assert Color.PASTEL_GREEN == 78
+
+    def test_downy(self):
+        assert Color.DOWNY == 79
+
+    def test_viking(self):
+        assert Color.VIKING == 80
+
+    def test_steel_blue(self):
+        assert Color.STEEL_BLUE == 81
+
+    def test_chartreuse(self):
+        assert Color.CHARTREUSE == 82
+
+    def test_screaming_green(self):
+        assert Color.SCREAMING_GREEN == 83
+
+    def test_sea_green_1(self):
+        assert Color.SEA_GREEN_1 == 84
+
+    def test_aquamarine_1(self):
+        assert Color.AQUAMARINE_1 == 85
+
+    def test_aquamarine_2(self):
+        assert Color.AQUAMARINE_2 == 86
+
+    def test_aquamarine(self):
+        assert Color.AQUAMARINE == 87
+
+    def test_dark_red(self):
+        assert Color.DARK_RED == 88
+
+    def test_fresh_eggplant(self):
+        assert Color.FRESH_EGGPLANT == 89
+
+    def test_dark_magenta(self):
+        assert Color.DARK_MAGENTA == 90
+
+    def test_purple_2(self):
+        assert Color.PURPLE_2 == 91
+
+    def test_electric_violet(self):
+        assert Color.ELECTRIC_VIOLET == 92
+
+    def test_purple_1(self):
+        assert Color.PURPLE_1 == 93
+
+    def test_brown(self):
+        assert Color.BROWN == 94
+
+    def test_copper_rose(self):
+        assert Color.COPPER_ROSE == 95
+
+    def test_strike_master(self):
+        assert Color.STRIKE_MASTER == 96
+
+    def test_deluge(self):
+        assert Color.DELUGE == 97
+
+    def test_medium_purple(self):
+        assert Color.MEDIUM_PURPLE == 98
+
+    def test_heliotrope(self):
+        assert Color.HELIOTROPE == 99
+
+    def test_olive_1(self):
+        assert Color.OLIVE_1 == 100
+
+    def test_clay_greek(self):
+        assert Color.CLAY_CREEK == 101
+
+    def test_gray_1(self):
+        assert Color.GRAY_1 == 102
+
+    def test_grey_1(self):
+        assert Color.GREY_1 == 102
+
+    def test_wild_blue_yonder(self):
+        assert Color.WILD_BLUE_YONDER == 103
+
+    def test_chetwode_blue(self):
+        assert Color.CHETWODE_BLUE == 104
+
+    def test_light_slate_blue(self):
+        assert Color.LIGHT_SLATE_BLUE == 105
+
+    def test_limeade_1(self):
+        assert Color.LIMEADE_1 == 106
+
+    def test_chelsea_cucumber(self):
+        assert Color.CHELSEA_CUCUMBER == 107
+
+    def test_bay_leaf(self):
+        assert Color.BAY_LEAF == 108
+
+    def test_gulf_stream(self):
+        assert Color.GULF_STREAM == 109
+
+    def test_polo_blue(self):
+        assert Color.POLO_BLUE == 110
+
+    def test_malibu_1(self):
+        assert Color.MALIBU_1 == 111
+
+    def test_pistachio(self):
+        assert Color.PISTACHIO == 112
+
+    def test_dark_olive_green(self):
+        assert Color.DARK_OLIVE_GREEN == 113
+
+    def test_feijoa(self):
+        assert Color.FEIJOA == 114
+
+    def test_vista_blue(self):
+        assert Color.VISTA_BLUE == 115
+
+    def test_bermuda(self):
+        assert Color.BERMUDA == 116
+
+    def test_anakiwa(self):
+        assert Color.ANAKIWA == 117
+
+    def test_charteuse_1(self):
+        assert Color.CHARTREUSE_1 == 118
+
+    def test_light_green(self):
+        assert Color.LIGHT_GREEN == 119
+
+    def test_mint_green(self):
+        assert Color.MINT_GREEN == 120
+
+    def test_pale_green_1(self):
+        assert Color.PALE_GREEN_1 == 121
+
+    def test_aqua_marine(self):
+        assert Color.AQUA_MARINE == 122
+
+    def test_anakiwa_1(self):
+        assert Color.ANAKIWA_1 == 123
+
+    def test_bright_red(self):
+        assert Color.BRIGHT_RED == 124
+
+    def test_flirt(self):
+        assert Color.FLIRT == 125
+
+    def test_medium_violet_red(self):
+        assert Color.MEDIUM_VIOLET_RED == 126
+
+    def test_magenta_1(self):
+        assert Color.MAGENTA_1 == 127
+
+    def test_dark_violet(self):
+        assert Color.DARK_VIOLET == 128
+
+    def test_purple_4(self):
+        assert Color.PURPLE_4 == 129
+
+    def test_rose_of_sharon(self):
+        assert Color.ROSE_OF_SHARON == 130
+
+    def test_indian_red(self):
+        assert Color.INDIAN_RED == 131
+
+    def test_tapestry(self):
+        assert Color.TAPESTRY == 132
+
+    def test_fuchsia_pink(self):
+        assert Color.FUCHSIA_PINK == 133
+
+    def test_medium_purple_1(self):
+        assert Color.MEDIUM_PURPLE_1 == 134
+
+    def test_heliotrope_1(self):
+        assert Color.HELIOTROPE_1 == 135
+
+    def test_pirate_gold(self):
+        assert Color.PIRATE_GOLD == 136
+
+    def test_muesli(self):
+        assert Color.MUESLI == 137
+
+    def test_pharlap(self):
+        assert Color.PHARLAP == 138
+
+    def test_bouquet(self):
+        assert Color.BOUQUET == 139
+
+    def test_lavender(self):
+        assert Color.LAVENDER == 140
+
+    def test_heliotrope_2(self):
+        assert Color.HELIOTROPE_2 == 141
+
+    def test_gold_1(self):
+        assert Color.GOLD_1 == 142
+
+    def test_olve_green(self):
+        assert Color.OLIVE_GREEN == 143
+
+    def test_hillary(self):
+        assert Color.HILLARY == 144
+
+    def test_silver_chalice(self):
+        assert Color.SILVER_CHALICE == 145
+
+    def test_wistful(self):
+        assert Color.WISTFUL == 146
+
+    def test_melrose(self):
+        assert Color.MELROSE == 147
+
+    def test_rio_grande(self):
+        assert Color.RIO_GRANDE == 148
+
+    def test_conifer(self):
+        assert Color.CONIFER == 149
+
+    def test_feijoa_1(self):
+        assert Color.FEIJOA_1 == 150
+
+    def test_pixie_green(self):
+        assert Color.PIXIE_GREEN == 151
+
+    def test_jungle_mist(self):
+        assert Color.JUNGLE_MIST == 152
+
+    def test_anakiwa_2(self):
+        assert Color.ANAKIWA_2 == 153
+
+    def test_lime_1(self):
+        assert Color.LIME_1 == 154
+
+    def test_green_yellow(self):
+        assert Color.GREEN_YELLOW == 155
+
+    def test_mint_green_1(self):
+        assert Color.MINT_GREEN_1 == 156
+
+    def test_dark_sea_green(self):
+        assert Color.DARK_SEA_GREEN == 157
+
+    def test_aero_blue(self):
+        assert Color.AERO_BLUE == 158
+
+    def test_french_pass(self):
+        assert Color.FRENCH_PASS == 159
+
+    def test_guardsman_red(self):
+        assert Color.GUARDSMAN_RED == 160
+
+    def test_razzmatazz(self):
+        assert Color.RAZZMATAZZ == 161
+
+    def test_hollywood_cerise(self):
+        assert Color.HOLLYWOOD_CERISE == 162
+
+    def test_hollywood_cerise_1(self):
+        assert Color.HOLLYWOOD_CERISE_1 == 163
+
+    def test_purple_pizzazz(self):
+        assert Color.PURPLE_PIZZAZZ == 164
+
+    def test_electric_violet_1(self):
+        assert Color.ELECTRIC_VIOLET_1 == 165
+
+    def test_tenn(self):
+        assert Color.TENN == 166
+
+    def test_roman(self):
+        assert Color.ROMAN == 167
+
+    def test_cranberry(self):
+        assert Color.CRANBERRY == 168
+
+    def test_hopbush(self):
+        assert Color.HOPBUSH == 169
+
+    def test_orchid(self):
+        assert Color.ORCHID == 170
+
+    def test_medium_orchid(self):
+        assert Color.MEDIUM_ORCHID == 171
+
+    def test_mango_tango(self):
+        assert Color.MANGO_TANGO == 172
+
+    def test_copperfield(self):
+        assert Color.COPPERFIELD == 173
+
+    def test_pink(self):
+        assert Color.PINK == 174
+
+    def test_cancan(self):
+        assert Color.CANCAN == 175
+
+    def test_light_orchid(self):
+        assert Color.LIGHT_ORCHID == 176
+
+    def test_heliotrope_3(self):
+        assert Color.HELIOTROPE_3 == 177
+
+    def test_corn(self):
+        assert Color.CORN == 178
+
+    def test_tacha(self):
+        assert Color.TACHA == 179
+
+    def test_tan(self):
+        assert Color.TAN == 180
+
+    def test_clam_shell(self):
+        assert Color.CLAM_SHELL == 181
+
+    def test_thistle(self):
+        assert Color.THISTLE == 182
+
+    def test_mauve(self):
+        assert Color.MAUVE == 183
+
+    def test_corn_1(self):
+        assert Color.CORN_1 == 184
+
+    def test_khaki(self):
+        assert Color.KHAKI == 185
+
+    def test_deco(self):
+        assert Color.DECO == 186
+
+    def test_green_mist(self):
+        assert Color.GREEN_MIST == 187
+
+    def test_alto(self):
+        assert Color.ALTO == 188
+
+    def test_FOG(self):
+        assert Color.FOG == 189
+
+    def test_chartreuse_yellow(self):
+        assert Color.CHARTREUSE_YELLOW == 190
+
+    def test_canary(self):
+        assert Color.CANARY == 191
+
+    def test_honeysuckle(self):
+        assert Color.HONEYSUCKLE == 192
+
+    def test_reef(self):
+        assert Color.REEF == 193
+
+    def test_snowy_mint(self):
+        assert Color.SNOWY_MINT == 194
+
+    def test_oyster_bay(self):
+        assert Color.OYSTER_BAY == 195
+
+    def test_rose(self):
+        assert Color.ROSE == 197
+
+    def test_deep_pink(self):
+        assert Color.DEEP_PINK == 198
+
+    def test_hollywood_cerise_2(self):
+        assert Color.HOLLYWOOD_CERISE_2 == 199
+
+    def test_purple_pizzazz_1(self):
+        assert Color.PURPLE_PIZZAZZ_1 == 200
+
+    def test_magenta(self):
+        assert Color.MAGENTA == 201
+
+    def test_blaze_orange(self):
+        assert Color.BLAZE_ORANGE == 202
+
+    def test_bitter_sweet(self):
+        assert Color.BITTER_SWEET == 203
+
+    def test_wild_watermelon(self):
+        assert Color.WILD_WATERMELON == 204
+
+    def test_hotpink(self):
+        assert Color.HOTPINK == 205
+
+    def test_hotpink_1(self):
+        assert Color.HOTPINK_1 == 206
+
+    def test_pink_flamingo(self):
+        assert Color.PINK_FLAMINGO == 207
+
+    def test_flush_orange(self):
+        assert Color.FLUSH_ORANGE == 208
+
+    def test_salmon(self):
+        assert Color.SALMON == 209
+
+    def test_vivid_tangerine(self):
+        assert Color.VIVID_TANGERINE == 210
+
+    def test_pink_salmon(self):
+        assert Color.PINK_SALMON == 211
+
+    def test_lavender_rose(self):
+        assert Color.LAVENDER_ROSE == 212
+
+    def test_blush_pink(self):
+        assert Color.BLUSH_PINK == 213
+
+    def test_yellow_sea(self):
+        assert Color.YELLOW_SEA == 214
+
+    def test_texas_rose(self):
+        assert Color.TEXAS_ROSE == 215
+
+    def test_hit_pink(self):
+        assert Color.HIT_PINK == 216
+
+    def test_sundown(self):
+        assert Color.SUNDOWN == 217
+
+    def test_cotton_candy(self):
+        assert Color.COTTON_CANDY == 218
+
+    def test_lavender_rose_1(self):
+        assert Color.LAVENDER_ROSE_1 == 219
+
+    def test_gold(self):
+        assert Color.GOLD == 220
+
+    def test_dandelion(self):
+        assert Color.DANDELION == 221
+
+    def test_grandis(self):
+        assert Color.GRANDIS == 222
+
+    def test_caramel(self):
+        assert Color.CARAMEL == 223
+
+    def test_cosmos(self):
+        assert Color.COSMOS == 224
+
+    def test_pink_lace(self):
+        assert Color.PINK_LACE == 225
+
+    def test_laser_lemon(self):
+        assert Color.LASER_LEMON == 227
+
+    def test_dolly(self):
+        assert Color.DOLLY == 228
+
+    def test_portafino(self):
+        assert Color.PORTAFINO == 229
+
+    def test_cumulus(self):
+        assert Color.CUMULUS == 230
+
+    def test_cod_gray(self):
+        assert Color.COD_GRAY == 232
+
+    def test_cod_gray_1(self):
+        assert Color.COD_GRAY_1 == 233
+
+    def test_cod_gray_2(self):
+        assert Color.COD_GRAY_2 == 234
+
+    def test_mine_shaft(self):
+        assert Color.MINE_SHAFT == 235
+
+    def test_mine_shaft_1(self):
+        assert Color.MINE_SHAFT_1 == 236
+
+    def test_mine_shaft_2(self):
+        assert Color.MINE_SHAFT_2 == 237
+
+    def test_tundura(self):
+        assert Color.TUNDORA == 238
+
+    def test_tundura_1(self):
+        assert Color.TUNDORA_1 == 239
+
+    def test_scorpion_1(self):
+        assert Color.SCORPION_1 == 240
+
+    def test_dove_gray(self):
+        assert Color.DOVE_GRAY == 241
+
+    def test_dove_gray_1(self):
+        assert Color.DOVE_GRAY_1 == 242
+
+    def test_boulder(self):
+        assert Color.BOULDER == 243
+
+    def test_gray_2(self):
+        assert Color.GRAY_2 == 245
+
+    def test_grey_2(self):
+        assert Color.GREY_2 == 245
+
+    def test_dusty_gray(self):
+        assert Color.DUSTY_GRAY == 246
+
+    def test_silver_chalice_1(self):
+        assert Color.SILVER_CHALICE_1 == 247
+
+    def test_silver_chalice_2(self):
+        assert Color.SILVER_CHALICE_2 == 248
+
+    def test_silver_chalice_3(self):
+        assert Color.SILVER_CHALICE_3 == 249
+
+    def test_silver_1(self):
+        assert Color.SILVER_1 == 250
+
+    def test_silver_2(self):
+        assert Color.SILVER_2 == 251
+
+    def test_alto_1(self):
+        assert Color.ALTO_1 == 252
+
+    def test_alto_2(self):
+        assert Color.ALTO_2 == 253
+
+    def test_mercury(self):
+        assert Color.MERCURY == 254
+
+    def test_gallery(self):
+        assert Color.GALLERY == 255
