+++ tests/highlight_test.py
@@ -1,765 +1,10 @@
+from termspark.painter.constants.color import Color
 from termspark.painter.constants.highlight import Highlight
 
 
 class TestHighlight:
-    def test_black(self):
-        assert Highlight.BLACK == 0
-
-    def test_maroon(self):
-        assert Highlight.MAROON == 1
-
-    def test_green(self):
-        assert Highlight.GREEN == 2
-
-    def test_olive(self):
-        assert Highlight.OLIVE == 3
-
-    def test_navy(self):
-        assert Highlight.NAVY == 4
-
-    def test_purple(self):
-        assert Highlight.PURPLE == 5
-
-    def test_teal_1(self):
-        assert Highlight.TEAL_1 == 6
-
-    def test_silver(self):
-        assert Highlight.SILVER == 7
-
-    def test_gray(self):
-        assert Highlight.GRAY == 8
-
-    def test_grey(self):
-        assert Highlight.GREY == 8
-
-    def test_red(self):
-        assert Highlight.RED == 9
-
-    def test_lime(self):
-        assert Highlight.LIME == 10
-
-    def test_yellow(self):
-        assert Highlight.YELLOW == 11
-
-    def test_blue(self):
-        assert Highlight.BLUE == 12
-
-    def test_fuchsia(self):
-        assert Highlight.FUCHSIA == 13
-
-    def test_aqua(self):
-        assert Highlight.AQUA == 14
-
-    def test_white(self):
-        assert Highlight.WHITE == 15
-
-    def test_navy_blue(self):
-        assert Highlight.NAVY_BLUE == 17
-
-    def test_dark_blue(self):
-        assert Highlight.DARK_BLUE == 18
-
-    def test_dark_blue_2(self):
-        assert Highlight.DARK_BLUE_2 == 19
-
-    def test_dark_blue_1(self):
-        assert Highlight.DARK_BLUE_1 == 20
-
-    def test_blue_1(self):
-        assert Highlight.BLUE_1 == 21
-
-    def test_dark_green(self):
-        assert Highlight.DARK_GREEN == 22
-
-    def test_blue_stone(self):
-        assert Highlight.BLUE_STONE == 23
-
-    def test_orient(self):
-        assert Highlight.ORIENT == 24
-
-    def test_endeavour(self):
-        assert Highlight.ENDEAVOUR == 25
-
-    def test_science_blue(self):
-        assert Highlight.SCIENCE_BLUE == 26
-
-    def test_blue_ribbon(self):
-        assert Highlight.BLUE_RIBBON == 27
-
-    def test_japanese_laurel(self):
-        assert Highlight.JAPANESE_LAUREL == 28
-
-    def test_deep_sea(self):
-        assert Highlight.DEEP_SEA == 29
-
-    def test_teal(self):
-        assert Highlight.TEAL == 30
-
-    def test_deep_cerulean(self):
-        assert Highlight.DEEP_CERULEAN == 31
-
-    def test_lochmara(self):
-        assert Highlight.LOCHMARA == 32
-
-    def test_azure_radiance(self):
-        assert Highlight.AZURE_RADIANCE == 33
-
-    def test_islamic_green(self):
-        assert Highlight.ISLAMIC_GREEN == 34
-
-    def test_spring_green(self):
-        assert Highlight.SPRING_GREEN == 35
-
-    def test_dark_cyan(self):
-        assert Highlight.DARK_CYAN == 36
-
-    def test_light_sea_green(self):
-        assert Highlight.LIGHT_SEA_GREEN == 37
-
-    def test_cerulean(self):
-        assert Highlight.CERULEAN == 38
-
-    def test_blue_bolt(self):
-        assert Highlight.BLUE_BOLT == 39
-
-    def test_electric_green(self):
-        assert Highlight.ELECTRIC_GREEN == 40
-
-    def test_malachite(self):
-        assert Highlight.MALACHITE == 41
-
-    def test_caribbean_green(self):
-        assert Highlight.CARIBBEAN_GREEN == 42
-
-    def test_cyan_1(self):
-        assert Highlight.CYAN_1 == 43
-
-    def test_dark_turquoise(self):
-        assert Highlight.DARK_TURQUOISE == 44
-
-    def test_vivid_sky_blue(self):
-        assert Highlight.VIVID_SKY_BLUE == 45
-
-    def test_electric_green_1(self):
-        assert Highlight.ELECTRIC_GREEN_1 == 46
-
-    def test_guppie_green(self):
-        assert Highlight.GUPPIE_GREEN == 47
-
-    def test_spring_green_1(self):
-        assert Highlight.SPRING_GREEN_1 == 48
-
-    def test_medium_spring_green(self):
-        assert Highlight.MEDIUM_SPRING_GREEN == 49
-
-    def test_sea_green(self):
-        assert Highlight.SEA_GREEN == 50
-
-    def test_cyan(self):
-        assert Highlight.CYAN == 51
-
-    def test_rosewood(self):
-        assert Highlight.ROSEWOOD == 52
-
-    def test_pompadour(self):
-        assert Highlight.POMPADOUR == 53
-
-    def test_pigment_indigo(self):
-        assert Highlight.PIGMENT_INDIGO == 54
-
-    def test_purple_3(self):
-        assert Highlight.PURPLE_3 == 55
-
-    def test_electic_violet(self):
-        assert Highlight.ELECTIC_VIOLET == 56
-
-    def test_blue_violet(self):
-        assert Highlight.BLUE_VIOLET == 57
-
-    def test_verdun_green(self):
-        assert Highlight.VERDUN_GREEN == 58
-
-    def test_scorpion(self):
-        assert Highlight.SCORPION == 59
-
-    def test_comet(self):
-        assert Highlight.COMET == 60
-
-    def test_scampi(self):
-        assert Highlight.SCAMPI == 61
-
-    def test_indigo(self):
-        assert Highlight.INDIGO == 62
-
-    def test_cornflower_blue_1(self):
-        assert Highlight.CORNFLOWER_BLUE_1 == 63
-
-    def test_limeade(self):
-        assert Highlight.LIMEADE == 64
-
-    def test_glade_green(self):
-        assert Highlight.GLADE_GREEN == 65
-
-    def test_juniper(self):
-        assert Highlight.JUNIPER == 66
-
-    def test_hippie_blue(self):
-        assert Highlight.HIPPIE_BLUE == 67
-
-    def test_havelock_blue(self):
-        assert Highlight.HAVELOCK_BLUE == 68
-
-    def test_cornflower_blue(self):
-        assert Highlight.CORNFLOWER_BLUE == 69
-
-    def test_limea(self):
-        assert Highlight.LIMEA == 70
-
-    def test_fern(self):
-        assert Highlight.FERN == 71
-
-    def test_silver_tree(self):
-        assert Highlight.SILVER_TREE == 72
-
-    def test_tradewind(self):
-        assert Highlight.TRADEWIND == 73
-
-    def test_shakespeare(self):
-        assert Highlight.SHAKESPEARE == 74
-
-    def test_malibu(self):
-        assert Highlight.MALIBU == 75
-
-    def test_bright_green(self):
-        assert Highlight.BRIGHT_GREEN == 76
-
-    def test_pale_green(self):
-        assert Highlight.PALE_GREEN == 77
-
-    def test_pastel_green(self):
-        assert Highlight.PASTEL_GREEN == 78
-
-    def test_downy(self):
-        assert Highlight.DOWNY == 79
-
-    def test_viking(self):
-        assert Highlight.VIKING == 80
-
-    def test_steel_blue(self):
-        assert Highlight.STEEL_BLUE == 81
-
-    def test_chartreuse(self):
-        assert Highlight.CHARTREUSE == 82
-
-    def test_screaming_green(self):
-        assert Highlight.SCREAMING_GREEN == 83
-
-    def test_sea_green_1(self):
-        assert Highlight.SEA_GREEN_1 == 84
-
-    def test_aquamarine_1(self):
-        assert Highlight.AQUAMARINE_1 == 85
-
-    def test_aquamarine_2(self):
-        assert Highlight.AQUAMARINE_2 == 86
-
-    def test_aquamarine(self):
-        assert Highlight.AQUAMARINE == 87
-
-    def test_dark_red(self):
-        assert Highlight.DARK_RED == 88
-
-    def test_fresh_eggplant(self):
-        assert Highlight.FRESH_EGGPLANT == 89
-
-    def test_dark_magenta(self):
-        assert Highlight.DARK_MAGENTA == 90
-
-    def test_purple_2(self):
-        assert Highlight.PURPLE_2 == 91
-
-    def test_electric_violet(self):
-        assert Highlight.ELECTRIC_VIOLET == 92
-
-    def test_purple_1(self):
-        assert Highlight.PURPLE_1 == 93
-
-    def test_brown(self):
-        assert Highlight.BROWN == 94
-
-    def test_copper_rose(self):
-        assert Highlight.COPPER_ROSE == 95
-
-    def test_strike_master(self):
-        assert Highlight.STRIKE_MASTER == 96
-
-    def test_deluge(self):
-        assert Highlight.DELUGE == 97
-
-    def test_medium_purple(self):
-        assert Highlight.MEDIUM_PURPLE == 98
-
-    def test_heliotrope(self):
-        assert Highlight.HELIOTROPE == 99
-
-    def test_olive_1(self):
-        assert Highlight.OLIVE_1 == 100
-
-    def test_clay_greek(self):
-        assert Highlight.CLAY_CREEK == 101
-
-    def test_gray_1(self):
-        assert Highlight.GRAY_1 == 102
-
-    def test_grey_1(self):
-        assert Highlight.GREY_1 == 102
-
-    def test_wild_blue_yonder(self):
-        assert Highlight.WILD_BLUE_YONDER == 103
-
-    def test_chetwode_blue(self):
-        assert Highlight.CHETWODE_BLUE == 104
-
-    def test_light_slate_blue(self):
-        assert Highlight.LIGHT_SLATE_BLUE == 105
-
-    def test_limeade_1(self):
-        assert Highlight.LIMEADE_1 == 106
-
-    def test_chelsea_cucumber(self):
-        assert Highlight.CHELSEA_CUCUMBER == 107
-
-    def test_bay_leaf(self):
-        assert Highlight.BAY_LEAF == 108
-
-    def test_gulf_stream(self):
-        assert Highlight.GULF_STREAM == 109
-
-    def test_polo_blue(self):
-        assert Highlight.POLO_BLUE == 110
-
-    def test_malibu_1(self):
-        assert Highlight.MALIBU_1 == 111
-
-    def test_pistachio(self):
-        assert Highlight.PISTACHIO == 112
-
-    def test_dark_olive_green(self):
-        assert Highlight.DARK_OLIVE_GREEN == 113
-
-    def test_feijoa(self):
-        assert Highlight.FEIJOA == 114
-
-    def test_vista_blue(self):
-        assert Highlight.VISTA_BLUE == 115
-
-    def test_bermuda(self):
-        assert Highlight.BERMUDA == 116
-
-    def test_anakiwa(self):
-        assert Highlight.ANAKIWA == 117
-
-    def test_charteuse_1(self):
-        assert Highlight.CHARTREUSE_1 == 118
-
-    def test_light_green(self):
-        assert Highlight.LIGHT_GREEN == 119
-
-    def test_mint_green(self):
-        assert Highlight.MINT_GREEN == 120
-
-    def test_pale_green_1(self):
-        assert Highlight.PALE_GREEN_1 == 121
-
-    def test_aqua_marine(self):
-        assert Highlight.AQUA_MARINE == 122
-
-    def test_anakiwa_1(self):
-        assert Highlight.ANAKIWA_1 == 123
-
-    def test_bright_red(self):
-        assert Highlight.BRIGHT_RED == 124
-
-    def test_flirt(self):
-        assert Highlight.FLIRT == 125
-
-    def test_medium_violet_red(self):
-        assert Highlight.MEDIUM_VIOLET_RED == 126
-
-    def test_magenta_1(self):
-        assert Highlight.MAGENTA_1 == 127
-
-    def test_dark_violet(self):
-        assert Highlight.DARK_VIOLET == 128
-
-    def test_purple_4(self):
-        assert Highlight.PURPLE_4 == 129
-
-    def test_rose_of_sharon(self):
-        assert Highlight.ROSE_OF_SHARON == 130
-
-    def test_indian_red(self):
-        assert Highlight.INDIAN_RED == 131
-
-    def test_tapestry(self):
-        assert Highlight.TAPESTRY == 132
-
-    def test_fuchsia_pink(self):
-        assert Highlight.FUCHSIA_PINK == 133
-
-    def test_medium_purple_1(self):
-        assert Highlight.MEDIUM_PURPLE_1 == 134
-
-    def test_heliotrope_1(self):
-        assert Highlight.HELIOTROPE_1 == 135
-
-    def test_pirate_gold(self):
-        assert Highlight.PIRATE_GOLD == 136
-
-    def test_muesli(self):
-        assert Highlight.MUESLI == 137
-
-    def test_pharlap(self):
-        assert Highlight.PHARLAP == 138
-
-    def test_bouquet(self):
-        assert Highlight.BOUQUET == 139
-
-    def test_lavender(self):
-        assert Highlight.LAVENDER == 140
-
-    def test_heliotrope_2(self):
-        assert Highlight.HELIOTROPE_2 == 141
-
-    def test_gold_1(self):
-        assert Highlight.GOLD_1 == 142
-
-    def test_olve_green(self):
-        assert Highlight.OLIVE_GREEN == 143
-
-    def test_hillary(self):
-        assert Highlight.HILLARY == 144
-
-    def test_silver_chalice(self):
-        assert Highlight.SILVER_CHALICE == 145
-
-    def test_wistful(self):
-        assert Highlight.WISTFUL == 146
-
-    def test_melrose(self):
-        assert Highlight.MELROSE == 147
-
-    def test_rio_grande(self):
-        assert Highlight.RIO_GRANDE == 148
-
-    def test_conifer(self):
-        assert Highlight.CONIFER == 149
-
-    def test_feijoa_1(self):
-        assert Highlight.FEIJOA_1 == 150
-
-    def test_pixie_green(self):
-        assert Highlight.PIXIE_GREEN == 151
-
-    def test_jungle_mist(self):
-        assert Highlight.JUNGLE_MIST == 152
-
-    def test_anakiwa_2(self):
-        assert Highlight.ANAKIWA_2 == 153
-
-    def test_lime_1(self):
-        assert Highlight.LIME_1 == 154
-
-    def test_green_yellow(self):
-        assert Highlight.GREEN_YELLOW == 155
-
-    def test_mint_green_1(self):
-        assert Highlight.MINT_GREEN_1 == 156
-
-    def test_dark_sea_green(self):
-        assert Highlight.DARK_SEA_GREEN == 157
-
-    def test_aero_blue(self):
-        assert Highlight.AERO_BLUE == 158
-
-    def test_french_pass(self):
-        assert Highlight.FRENCH_PASS == 159
-
-    def test_guardsman_red(self):
-        assert Highlight.GUARDSMAN_RED == 160
-
-    def test_razzmatazz(self):
-        assert Highlight.RAZZMATAZZ == 161
-
-    def test_hollywood_cerise(self):
-        assert Highlight.HOLLYWOOD_CERISE == 162
-
-    def test_hollywood_cerise_1(self):
-        assert Highlight.HOLLYWOOD_CERISE_1 == 163
-
-    def test_purple_pizzazz(self):
-        assert Highlight.PURPLE_PIZZAZZ == 164
-
-    def test_electric_violet_1(self):
-        assert Highlight.ELECTRIC_VIOLET_1 == 165
-
-    def test_tenn(self):
-        assert Highlight.TENN == 166
-
-    def test_roman(self):
-        assert Highlight.ROMAN == 167
-
-    def test_cranberry(self):
-        assert Highlight.CRANBERRY == 168
-
-    def test_hopbush(self):
-        assert Highlight.HOPBUSH == 169
-
-    def test_orchid(self):
-        assert Highlight.ORCHID == 170
-
-    def test_medium_orchid(self):
-        assert Highlight.MEDIUM_ORCHID == 171
-
-    def test_mango_tango(self):
-        assert Highlight.MANGO_TANGO == 172
-
-    def test_copperfield(self):
-        assert Highlight.COPPERFIELD == 173
-
-    def test_pink(self):
-        assert Highlight.PINK == 174
-
-    def test_cancan(self):
-        assert Highlight.CANCAN == 175
-
-    def test_light_orchid(self):
-        assert Highlight.LIGHT_ORCHID == 176
-
-    def test_heliotrope_3(self):
-        assert Highlight.HELIOTROPE_3 == 177
-
-    def test_corn(self):
-        assert Highlight.CORN == 178
-
-    def test_tacha(self):
-        assert Highlight.TACHA == 179
-
-    def test_tan(self):
-        assert Highlight.TAN == 180
-
-    def test_clam_shell(self):
-        assert Highlight.CLAM_SHELL == 181
-
-    def test_thistle(self):
-        assert Highlight.THISTLE == 182
-
-    def test_mauve(self):
-        assert Highlight.MAUVE == 183
-
-    def test_corn_1(self):
-        assert Highlight.CORN_1 == 184
-
-    def test_khaki(self):
-        assert Highlight.KHAKI == 185
-
-    def test_deco(self):
-        assert Highlight.DECO == 186
-
-    def test_green_mist(self):
-        assert Highlight.GREEN_MIST == 187
-
-    def test_alto(self):
-        assert Highlight.ALTO == 188
-
-    def test_FOG(self):
-        assert Highlight.FOG == 189
-
-    def test_chartreuse_yellow(self):
-        assert Highlight.CHARTREUSE_YELLOW == 190
-
-    def test_canary(self):
-        assert Highlight.CANARY == 191
-
-    def test_honeysuckle(self):
-        assert Highlight.HONEYSUCKLE == 192
-
-    def test_reef(self):
-        assert Highlight.REEF == 193
-
-    def test_snowy_mint(self):
-        assert Highlight.SNOWY_MINT == 194
-
-    def test_oyster_bay(self):
-        assert Highlight.OYSTER_BAY == 195
-
-    def test_rose(self):
-        assert Highlight.ROSE == 197
-
-    def test_deep_pink(self):
-        assert Highlight.DEEP_PINK == 198
-
-    def test_hollywood_cerise_2(self):
-        assert Highlight.HOLLYWOOD_CERISE_2 == 199
-
-    def test_purple_pizzazz_1(self):
-        assert Highlight.PURPLE_PIZZAZZ_1 == 200
-
-    def test_magenta(self):
-        assert Highlight.MAGENTA == 201
-
-    def test_blaze_orange(self):
-        assert Highlight.BLAZE_ORANGE == 202
-
-    def test_bitter_sweet(self):
-        assert Highlight.BITTER_SWEET == 203
-
-    def test_wild_watermelon(self):
-        assert Highlight.WILD_WATERMELON == 204
-
-    def test_hotpink(self):
-        assert Highlight.HOTPINK == 205
-
-    def test_hotpink_1(self):
-        assert Highlight.HOTPINK_1 == 206
-
-    def test_pink_flamingo(self):
-        assert Highlight.PINK_FLAMINGO == 207
-
-    def test_flush_orange(self):
-        assert Highlight.FLUSH_ORANGE == 208
-
-    def test_salmon(self):
-        assert Highlight.SALMON == 209
-
-    def test_vivid_tangerine(self):
-        assert Highlight.VIVID_TANGERINE == 210
-
-    def test_pink_salmon(self):
-        assert Highlight.PINK_SALMON == 211
-
-    def test_lavender_rose(self):
-        assert Highlight.LAVENDER_ROSE == 212
-
-    def test_blush_pink(self):
-        assert Highlight.BLUSH_PINK == 213
-
-    def test_yellow_sea(self):
-        assert Highlight.YELLOW_SEA == 214
-
-    def test_texas_rose(self):
-        assert Highlight.TEXAS_ROSE == 215
-
-    def test_hit_pink(self):
-        assert Highlight.HIT_PINK == 216
-
-    def test_sundown(self):
-        assert Highlight.SUNDOWN == 217
-
-    def test_cotton_candy(self):
-        assert Highlight.COTTON_CANDY == 218
-
-    def test_lavender_rose_1(self):
-        assert Highlight.LAVENDER_ROSE_1 == 219
-
-    def test_gold(self):
-        assert Highlight.GOLD == 220
-
-    def test_dandelion(self):
-        assert Highlight.DANDELION == 221
-
-    def test_grandis(self):
-        assert Highlight.GRANDIS == 222
-
-    def test_caramel(self):
-        assert Highlight.CARAMEL == 223
-
-    def test_cosmos(self):
-        assert Highlight.COSMOS == 224
-
-    def test_pink_lace(self):
-        assert Highlight.PINK_LACE == 225
-
-    def test_laser_lemon(self):
-        assert Highlight.LASER_LEMON == 227
-
-    def test_dolly(self):
-        assert Highlight.DOLLY == 228
-
-    def test_portafino(self):
-        assert Highlight.PORTAFINO == 229
-
-    def test_cumulus(self):
-        assert Highlight.CUMULUS == 230
-
-    def test_cod_gray(self):
-        assert Highlight.COD_GRAY == 232
-
-    def test_cod_gray_1(self):
-        assert Highlight.COD_GRAY_1 == 233
-
-    def test_cod_gray_2(self):
-        assert Highlight.COD_GRAY_2 == 234
-
-    def test_mine_shaft(self):
-        assert Highlight.MINE_SHAFT == 235
-
-    def test_mine_shaft_1(self):
-        assert Highlight.MINE_SHAFT_1 == 236
-
-    def test_mine_shaft_2(self):
-        assert Highlight.MINE_SHAFT_2 == 237
-
-    def test_tundura(self):
-        assert Highlight.TUNDORA == 238
-
-    def test_tundura_1(self):
-        assert Highlight.TUNDORA_1 == 239
-
-    def test_scorpion_1(self):
-        assert Highlight.SCORPION_1 == 240
-
-    def test_dove_gray(self):
-        assert Highlight.DOVE_GRAY == 241
-
-    def test_dove_gray_1(self):
-        assert Highlight.DOVE_GRAY_1 == 242
-
-    def test_boulder(self):
-        assert Highlight.BOULDER == 243
-
-    def test_gray_2(self):
-        assert Highlight.GRAY_2 == 245
-
-    def test_grey_2(self):
-        assert Highlight.GREY_2 == 245
-
-    def test_dusty_gray(self):
-        assert Highlight.DUSTY_GRAY == 246
-
-    def test_silver_chalice_1(self):
-        assert Highlight.SILVER_CHALICE_1 == 247
-
-    def test_silver_chalice_2(self):
-        assert Highlight.SILVER_CHALICE_2 == 248
-
-    def test_silver_chalice_3(self):
-        assert Highlight.SILVER_CHALICE_3 == 249
-
-    def test_silver_1(self):
-        assert Highlight.SILVER_1 == 250
-
-    def test_silver_2(self):
-        assert Highlight.SILVER_2 == 251
-
-    def test_alto_1(self):
-        assert Highlight.ALTO_1 == 252
-
-    def test_alto_2(self):
-        assert Highlight.ALTO_2 == 253
-
-    def test_mercury(self):
-        assert Highlight.MERCURY == 254
-
-    def test_gallery(self):
-        assert Highlight.GALLERY == 255
+    def test_prefix(self):
+        assert Highlight.PREFIX == "\x1b[48;5;"
+
+    def test_inherited_from_color(self):
+        assert issubclass(Highlight, Color)
